Introduce simulation mode

+++ handle_prometheus.py
@@ -3,6 +3,7 @@ import requests
 from ruamel import yaml
 import handle_docker as dock
 import shutil,os
+import pk_config
 
 alerts = {}
 
@@ -54,6 +55,8 @@ def evaluate_data_queries_for_nodes(endpoint,policy):
     for param,query in policy['data']['queries'].iteritems():
       try:
         if target_str is not None and target_str.find(param) != -1:
+          if pk_config.simulate():
+            continue
           response = requests.get(endpoint+"/api/v1/query?query="+query).json()
           log.debug('Prometheus response query "{0}":{1}'.format(query,response))
           val = extract_value_from_prometheus_response(query,response,dict())
@@ -77,6 +80,8 @@ def evaluate_data_queries_for_a_service(endpoint,policy,servicename):
     for param,query in policy['data']['queries'].iteritems():
       try:
         if target_str is not None and target_str.find(param) != -1:
+          if pk_config.simulate():
+            continue
           response = requests.get(endpoint+"/api/v1/query?query="+query).json()
           log.debug('Prometheus response query "{0}":{1}'.format(query,response))
           val = extract_value_from_prometheus_response(query,response,dict())
@@ -91,28 +96,30 @@ def evaluate_data_queries_for_a_service(endpoint,policy,servicename):
 def add_exporters_to_prometheus_config(policy, template_file, config_file):
   log=logging.getLogger('pk_prometheus')
   try:
-    with open(template_file,'r') as f:
-      config_content = yaml.round_trip_load(f)
-      if 'scrape_configs' not in config_content:
-        config_content['scrape_configs']=[]
-      #Find proper scrape_config or create
-      scrape_config = [ x for x in config_content['scrape_configs'] 
-                        if x.get('job_name','')=='micado' and 'static_configs' in x ]
-      if not scrape_config:
-        config_content['scrape_configs'].append({'job_name': 'micado','static_configs':[]})
-        scrape_config = [ x for x in config_content['scrape_configs']
-                        if x.get('job_name','')=='micado' and 'static_configs' in x ][0]
-      else:
-        scrape_config = scrape_config[0]
-      #Find proper static_config or create
+    config_content = dict()
+    if not pk_config.simulate():
+      with open(template_file,'r') as f:
+        config_content = yaml.round_trip_load(f)
+    if 'scrape_configs' not in config_content:
+      config_content['scrape_configs']=[]
+    #Find proper scrape_config or create
+    scrape_config = [ x for x in config_content['scrape_configs'] 
+		      if x.get('job_name','')=='micado' and 'static_configs' in x ]
+    if not scrape_config:
+      config_content['scrape_configs'].append({'job_name': 'micado','static_configs':[]})
+      scrape_config = [ x for x in config_content['scrape_configs']
+		      if x.get('job_name','')=='micado' and 'static_configs' in x ][0]
+    else:
+      scrape_config = scrape_config[0]
+    #Find proper static_config or create
+    static_config = [ x for x in scrape_config['static_configs']
+		    if 'targets' in x.keys() ]
+    if not static_config:
+      scrape_config['static_configs'].append({'targets': []})
       static_config = [ x for x in scrape_config['static_configs']
-                      if 'targets' in x.keys() ]
-      if not static_config:
-        scrape_config['static_configs'].append({'targets': []})
-        static_config = [ x for x in scrape_config['static_configs']
-                        if 'targets' in x.keys() ][0]
-      else:
-        static_config = static_config[0]
+		      if 'targets' in x.keys() ][0]
+    else:
+      static_config = static_config[0]
 
     config_changed = False 
     for exporter_endpoint in policy.get('data',dict()).get('sources',dict()):
@@ -123,7 +130,7 @@ def add_exporters_to_prometheus_config(policy, template_file, config_file):
       else:
         log.info('(C) => exporter "{0}" skipped, already part of config'.format(exporter_endpoint))
 
-    if config_changed:
+    if config_changed and not pk_config.simulate():
       with open(config_file, 'w') as outfile:
         yaml.round_trip_dump(config_content, outfile, default_flow_style=False)
 
@@ -133,7 +140,8 @@ def add_exporters_to_prometheus_config(policy, template_file, config_file):
   return
 
 def remove_exporters_from_prometheus_config(template_file, config_file):
-  shutil.copyfile(template_file, config_file)
+  if not pk_config.simulate():
+    shutil.copyfile(template_file, config_file)
 
 def attach_prometheus_to_exporters_network(policy,swarm_endpoint):
   log=logging.getLogger('pk_prometheus')
@@ -142,6 +150,8 @@ def attach_prometheus_to_exporters_network(policy,swarm_endpoint):
       exporter_name=exporter_endpoint.split(':')[0]
       if '.' not in exporter_name:
         log.info('(C) => attaching prometheus to network of exporter "{0}"'.format(exporter_endpoint))
+        if pk_config.simulate():
+          continue
         exporter_netid = dock.query_service_network(swarm_endpoint,policy['stack'],exporter_name)
         if exporter_netid:
           dock.attach_container_to_network(swarm_endpoint, 'prometheus', exporter_netid)
@@ -155,6 +165,8 @@ def detach_prometheus_from_exporters_network(policy,swarm_endpoint):
       exporter_name=exporter_endpoint.split(':')[0]
       if '.' not in exporter_name:
         log.info('(C) => detaching prometheus from network of exporter "{0}"'.format(exporter_endpoint))
+        if pk_config.simulate():
+          continue
         exporter_netid = dock.query_service_network(swarm_endpoint,policy['stack'],exporter_name)
         if exporter_netid:
           dock.detach_container_from_network(swarm_endpoint, 'prometheus', exporter_netid)
@@ -164,7 +176,8 @@ def detach_prometheus_from_exporters_network(policy,swarm_endpoint):
 def notify_to_reload_config(endpoint):
   log=logging.getLogger('pk_prometheus')
   try:
-    requests.post(endpoint+"/-/reload")
+    if not pk_config.simulate():
+      requests.post(endpoint+"/-/reload")
     log.info('(C) Notification to reload config sent to Prometheus.')
   except Exception:
     log.exception('Sending config reload notification to Prometheus failed:')
@@ -182,8 +195,9 @@ def deploy_alerts_under_prometheus(rules_directory,alerts,stack):
     for alert in alerts:
       content['groups'][0]['rules'].append(dict(alert))
     rule_file=os.path.join(rules_directory,stack+'.rules')
-    with open(rule_file, 'w') as outfile:
-      yaml.round_trip_dump(content, outfile, default_flow_style=False)
+    if not pk_config.simulate():
+      with open(rule_file, 'w') as outfile:
+        yaml.round_trip_dump(content, outfile, default_flow_style=False)
   except Exception:
     log.exception('Deploying alerts under Prometheus failed:')
   return
@@ -194,7 +208,8 @@ def remove_alerts_under_prometheus(rules_directory,alerts,stack):
   log=logging.getLogger('pk_prometheus')
   try:
     rule_file=os.path.join(rules_directory,stack+'.rules')
-    os.remove(rule_file)
+    if not pk_config.simulate():
+      os.remove(rule_file)
   except Exception:
     log.exception('Removing alerts under Prometheus failed:')
   return
